Fix type errors from typed enum fields in IR entity ID structs
Update all sites that construct or decompose IRBlockId, IRInstructionId,
and IRStructureId to use proper static_cast<> between typed enums
(ir::BlockKind, ir::OpCode, ir::StructureKind) and uint8_t, matching
the struct field type changes in Types.h.

+++ lib/IR/Block.cpp
@@ -27,7 +27,7 @@ EntityId IRBlock::id(void) const {
   IRBlockId bid;
   bid.fragment_id = impl->fragment_id;
   bid.offset = impl->offset;
-  bid.block_kind = static_cast<uint8_t>(kind());
+  bid.block_kind = kind();
   return EntityId(bid);
 }
